+++ examples/resources/morpheus_textarea_option_type/import.sh
@@ -1,0 +1,1 @@
+terraform import morpheus_textarea_option_type.tf_example_textarea_option_type 1